refactored and made code more modular for readability and better maintenance
