+++ .gitignore
@@ -185,3 +185,4 @@ cython_debug/
 .cursorindexingignore
 
 CLAUDE.md
+.claudeignore
